adding both logins to login.js
